requirements.txt
pytest==8.3.3
PyYAML==6.0.2
termcolor==2.5.0